Make the tests PSR-4 compliant and load them only in the PHPUnit bootstrap
