+++ tests/bootstrap.php
@@ -19,4 +19,6 @@ if ( !is_readable( __DIR__ . '/../vendor/autoload.php' ) ) {
 	die( 'You need to install this package with Composer before you can run the tests' );
 }
 
-require_once( __DIR__ . '/../vendor/autoload.php' );
+$autoLoader = require_once( __DIR__ . '/../vendor/autoload.php' );
+
+$autoLoader->addPsr4( 'Diff\\Tests\\', __DIR__ . '/phpunit/' );
