+++ tests/phpunit/DiffOp/Diff/DiffAsOpTest.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace Diff\Tests;
+namespace Diff\Tests\DiffOp\Diff;
 
 use Diff\Diff;
 use Diff\DiffOpAdd;
 use Diff\DiffOpRemove;
+use Diff\Tests\DiffOp\DiffOpTest;
 
 /**
  * @covers Diff\Diff
