--- tests/phpunit/differ/OrderedListDifferTest.php
@@ -1,302 +1,0 @@
-<?php
-
-namespace Diff\Tests;
-
-use Diff\Comparer\CallbackComparer;
-use Diff\Differ;
-use Diff\DiffOpAdd;
-use Diff\DiffOpRemove;
-use Diff\OrderedListDiffer;
-
-/**
- * @covers Diff\OrderedListDiffer
- *
- * @since 0.9
- *
- * @group Diff
- * @group Differ
- *
- * @licence GNU GPL v2+
- * @author Jeroen De Dauw < jeroendedauw@gmail.com >
- * @author Tobias Gritschacher < tobias.gritschacher@wikimedia.de >
- */
-class OrderedListDifferTest extends DiffTestCase {
-
-	/**
-	 * Returns those that both work for native and strict mode.
-	 */
-	protected function getCommonArgLists() {
-		$argLists = array();
-
-		$old = array();
-		$new = array();
-		$expected = array();
-
-		$argLists[] = array( $old, $new, $expected,
-			'There should be no difference between empty arrays' );
-
-
-		$old = array( 42 );
-		$new = array( 42 );
-		$expected = array();
-
-		$argLists[] = array( $old, $new, $expected,
-			'There should be no difference between arrays with the same element' );
-
-
-		$old = array( 42, 'ohi', 4.2, false );
-		$new = array( 42, 'ohi', 4.2, false );
-		$expected = array();
-
-		$argLists[] = array( $old, $new, $expected,
-			'There should be no difference between arrays with the same elements' );
-
-
-		$old = array( 42, 'ohi', 4.2, false );
-		$new = array( false, 4.2, 'ohi', 42 );
-		$expected = array(
-			new DiffOpAdd( false ),
-			new DiffOpAdd( 4.2 ),
-			new DiffOpAdd( 'ohi' ),
-			new DiffOpAdd( 42 ),
-			new DiffOpRemove( 42 ),
-			new DiffOpRemove( 'ohi' ),
-			new DiffOpRemove( 4.2 ),
-			new DiffOpRemove( false )
-		);
-
-		$argLists[] = array( $old, $new, $expected,
-			'Changing the order of all four elements should result in four add operations and four remove operations' );
-
-
-		$old = array( 42, 'ohi', 4.2, false );
-		$new = array( 4.2, 'ohi', 42, false );
-		$expected = array(
-			new DiffOpAdd( 4.2 ),
-			new DiffOpAdd( 42 ),
-			new DiffOpRemove( 42 ),
-			new DiffOpRemove( 4.2 ),
-		);
-
-		$argLists[] = array( $old, $new, $expected,
-			'Changing the order of two of four elements should result in two add operations and two remove operations' );
-
-
-		$old = array();
-		$new = array( 42 );
-		$expected = array( new DiffOpAdd( 42 ) );
-
-		$argLists[] = array( $old, $new, $expected,
-			'An array with a single element should be an add operation different from an empty array' );
-
-
-		$old = array( 42 );
-		$new = array();
-		$expected = array( new DiffOpRemove( 42 ) );
-
-		$argLists[] = array( $old, $new, $expected,
-			'An empty array should be a remove operation different from an array with one element' );
-
-
-		$old = array( 1 );
-		$new = array( 2 );
-		$expected = array( new DiffOpRemove( 1 ), new DiffOpAdd( 2 ) );
-
-		$argLists[] = array( $old, $new, $expected,
-			'Two arrays with a single different element should differ by an add and a remove op' );
-
-
-		$old = array( 9001, 42, 1, 0 );
-		$new = array( 9001, 42, 2, 0 );
-		$expected = array( new DiffOpRemove( 1 ), new DiffOpAdd( 2 ) );
-
-		$argLists[] = array( $old, $new, $expected,
-			'Two arrays with a single different element should differ by an add and a remove op
-			 when the order of the identical elements stays the same' );
-
-
-		$old = array( 'a', 'b', 'c' );
-		$new = array( 'c', 'b', 'a', 'd' );
-		$expected = array(
-			new DiffOpRemove( 'a' ),
-			new DiffOpRemove( 'c' ),
-			new DiffOpAdd( 'c' ),
-			new DiffOpAdd( 'a' ),
-			new DiffOpAdd( 'd' )
-		);
-
-		$argLists[] = array( $old, $new, $expected,
-			'Changing the position of two elements and adding one new element should result
-			in two remove ops and three add ops' );
-
-
-		$old = array( 'a', 'b', 'c', 'd' );
-		$new = array( 'b', 'a', 'c' );
-		$expected = array(
-			new DiffOpRemove( 'a' ),
-			new DiffOpRemove( 'b' ),
-			new DiffOpRemove( 'd' ),
-			new DiffOpAdd( 'b' ),
-			new DiffOpAdd( 'a' )
-		);
-
-		$argLists[] = array( $old, $new, $expected,
-			'Changing the position of two elements and removing the last element should result
-			in three remove ops and two add ops' );
-
-
-		$old = array( 'a', 'b', 'c' );
-		$new = array( 'b', 'c' );
-		$expected = array(
-			new DiffOpRemove( 'a' ),
-			new DiffOpRemove( 'b' ),
-			new DiffOpRemove( 'c' ),
-			new DiffOpAdd( 'b' ),
-			new DiffOpAdd( 'c' )
-		);
-
-		$argLists[] = array( $old, $new, $expected,
-			'Removing the first element results in remove ops for all elements and add ops for the remaining elements,
-			 because the position of all remaining elements has changed' );
-
-		return $argLists;
-	}
-
-	public function toDiffProvider() {
-		$argLists = $this->getCommonArgLists();
-
-		$old = array( 42, 42 );
-		$new = array( 42 );
-		$expected = array( new DiffOpRemove( 42 ) );
-
-		$argLists[] = array( $old, $new, $expected,
-			'[42, 42] to [42] should [rem(42)]' );
-
-
-		$old = array( 42 );
-		$new = array( 42, 42 );
-		$expected = array( new DiffOpAdd( 42 ) );
-
-		$argLists[] = array( $old, $new, $expected,
-			'[42] to [42, 42] should [add(42)]' );
-
-
-		$old = array( '42' );
-		$new = array( 42 );
-		$expected = array( new DiffOpRemove( '42' ), new DiffOpAdd( 42 ) );
-
-		$argLists[] = array( $old, $new, $expected,
-			'["42"] to [42] should [rem("42"), add(42)]' );
-
-
-		$old = array( array( 1 ) );
-		$new = array( array( 2 ) );
-		$expected = array( new DiffOpRemove( array( 1 ) ), new DiffOpAdd( array( 2 ) ) );
-
-		$argLists[] = array( $old, $new, $expected,
-			'[[1]] to [[2]] should [rem([1]), add([2])]' );
-
-
-		$old = array( array( 2 ) );
-		$new = array( array( 2 ) );
-		$expected = array();
-
-		$argLists[] = array( $old, $new, $expected,
-			'[[2]] to [[2]] should result in an empty diff' );
-
-		// test "soft" object comparison
-		$obj1 = new \stdClass();
-		$obj2 = new \stdClass();
-		$objX = new \stdClass();
-
-		$obj1->test = "Test";
-		$obj2->test = "Test";
-		$objX->xest = "Test";
-
-		$old = array( $obj1 );
-		$new = array( $obj2 );
-		$expected = array( );
-
-		$argLists[] = array( $old, $new, $expected,
-			'Two arrays containing equivalent objects should result in an empty diff' );
-
-		$old = array( $obj1 );
-		$new = array( $objX );
-		$expected = array( new DiffOpRemove( $obj1 ), new DiffOpAdd( $objX )  );
-
-		$argLists[] = array( $old, $new, $expected,
-			'Two arrays containing different objects of the same type should result in an add and a remove op.' );
-
-		return $argLists;
-	}
-
-	/**
-	 * @dataProvider toDiffProvider
-	 */
-	public function testDoDiff( $old, $new, $expected, $message = '' ) {
-		$callback = function( $foo, $bar ) {
-			return is_object( $foo ) ? $foo == $bar : $foo === $bar;
-		};
-
-		$this->doTestDiff(
-			new OrderedListDiffer( new CallbackComparer( $callback ) ),
-			$old,
-			$new,
-			$expected,
-			$message
-		);
-	}
-
-	protected function doTestDiff( Differ $differ, $old, $new, $expected, $message ) {
-		$actual = $differ->doDiff( $old, $new );
-
-		$this->assertArrayEquals( $expected, $actual, false, false, $message );
-	}
-
-	public function testCallbackComparisonReturningFalse() {
-		$differ = new OrderedListDiffer( new CallbackComparer( function( $foo, $bar ) {
-			return false;
-		} ) );
-
-		$actual = $differ->doDiff( array( 1, '2' ), array( 1, '2', 'foo' ) );
-
-		$expected = array(
-			new DiffOpAdd( 1 ),
-			new DiffOpAdd( '2' ),
-			new DiffOpAdd( 'foo' ),
-			new DiffOpRemove( 1 ),
-			new DiffOpRemove( '2' ),
-		);
-
-		$this->assertArrayEquals(
-			$expected, $actual, false, false,
-			'All elements should be removed and added when comparison callback always returns false'
-		);
-	}
-
-	public function testCallbackComparisonReturningTrue() {
-		$differ = new OrderedListDiffer( new CallbackComparer( function( $foo, $bar ) {
-			return true;
-		} ) );
-
-		$actual = $differ->doDiff( array( 1, '2', 'baz' ), array( 1, 'foo', '2' ) );
-
-		$expected = array();
-
-		$this->assertArrayEquals(
-			$expected, $actual, false, false,
-			'No elements should be removed or added when comparison callback always returns true'
-		);
-	}
-
-	public function testCallbackComparisonReturningNyanCat() {
-		$differ = new OrderedListDiffer( new CallbackComparer( function( $foo, $bar ) {
-			return '~=[,,_,,]:3';
-		} ) );
-
-		$this->setExpectedException( 'RuntimeException' );
-
-		$differ->doDiff( array( 1, '2', 'baz' ), array( 1, 'foo', '2' ) );
-	}
-
-}
